Rename Routes to Router and let IonApp be included at App level
